refactor: relocate model files
Moved campaign, campaign-category, and donation model files from shared prisma schemas to local app directories. Updated all relevant imports to reference the new locations, improving code organization and reducing cross-package dependencies.
