+++ apps/campaign/src/campaign-category/dtos/response/campaign-category.response.ts
@@ -1,4 +1,4 @@
-import { CampaignCategory } from "@libs/databases/prisma/schemas/models/campaign-category.model"
+import { CampaignCategory } from "apps/campaign/src/campaign-category/models/campaign-category.model"
 import { Field, Int, ObjectType } from "@nestjs/graphql"
 
 @ObjectType("CampaignCategoriesResponse")
